Reuse snapshot maps (#1104)
Changes the mode of operation where we created new maps for each
snapshot and destroyed the maps when deleting the snapshot to a mode
where we have 2 maps and use one for writing and one for storing.

Notes:
- responsibility for updating JobExecutionRecord, which contains 
changing data from JobRecord moved from JobCoordinationService to 
MasterContext

- MasterContext contains final field with JobExecutionRecord. It mutates 
it (therefore some fields in it are volatile) and writes it out after 
each change. Since it can be mutated in parallel and parallel writes can 
be reordered, we use timestamp to ensure that an older update doesn't 
overwrite a newer one.

- One snapshot is stored in one map (previously, each vertex used its 
own map). The chunk key now contains vertex name and snapshotId. 
SnapshotId was added to it to distinguish zombie writes, should they 
ever happen, or if writing after a failed `IMap.clear`. When reading, 
unknown vertices and snapshotIds are logged and ignored.

- Which data map is used for new snapshot is communicated to members
using SnapshotOp. Because target member must know this before it starts
writing, we've removed the (rare) possibility that snapshot is written
before the operation is received.

- SnapshotRecord and SnapshotRepository classes were removed

Fixes #1115