Fix Kafka source responsiveness when broker down (#1429)
When the broker is down, the `poll(long)` and `commitSync()` calls don't
return. The deprecated `poll(long)` call was replaced with
`poll(Duration)` which should not have this problem. `commitSync` is
called only if some records were returned by the `poll` - this doesn't
prevent the issue completely if the broker fails between the `poll` and
`commitSync` calls, but if it fails while `poll` is blocking, it will
prevent stucking.

There is another possibility of getting stuck when fetching metadata in
`assignPartitions`, which I don't know how to fix.