Fix outbox blocking in snapshot phase 2 (#1883)
This happened:

- in `ProcessorTasklet.processInbox` we discovered that a new phase-2 begun

- there was an unfinished item in the outbox, so we blocked the outbox and
didn't begin the phase-2 in order to ensure that the unfinished item is finished

- however before that, the processor had returned from `process()` with an
empty inbox and an unfinished item. Therefore the process went to `fillInbox`
and found a WM there. So the state proceeded to PROCESS_WATERMARK, but the
outbox was still blocked.

- `tryProcessWatermark` tried to emit the `currentTraverser` first, but
couldn't because the outbox was blocked. And there it remained stuck, trying
that forever.

Returning `true` from `tryProcess()` or returning an empty inbox and having an
unfinished item in the outbox is allowed. So we changed the contract of
`snapshotCommitFinish` that it must not add items to outbox and if there's a
phase-2, we just execute that method and return to where we left.

Fixes #1881